requirements.txt
dj-database-url
django==1.9
django-userena
django-extensions
djangorestframework
gunicorn
psycopg2
whitenoise